Tidy up bar charts.
