+++ Sources/SwiftUICharts/BarChart/Models/Protocols/BarChartProtocolsExtensions.swift
@@ -161,32 +161,3 @@ extension CTMultiBarChartDataProtocol {
 //        }
     }
 }
-
-// MARK: - Bar Size and Position
-internal struct BarLayout {
-    internal static func barFrame(_ section: CGSize, _ widthFactor: CGFloat, _ value: Double, _ maxValue: Double) -> CGSize {
-        return CGSize(width: barWidth(section.width, widthFactor),
-                      height: barHeight(section.height, value, maxValue))
-    }
-    
-    internal static func barWidth(_ totalWidth: CGFloat, _ widthFactor: CGFloat) -> CGFloat {
-        return totalWidth * widthFactor
-    }
-    
-    internal static func barHeight(_ totalHeight: CGFloat, _ value: Double, _ maxValue: Double) -> CGFloat {
-        return totalHeight * divideByZeroProtection(CGFloat.self, value, maxValue)
-    }
-    
-    internal static func barOffset(_ section: CGSize, _ widthFactor: CGFloat, _ value: Double, _ maxValue: Double) -> CGSize {
-        return CGSize(width: barXOffset(section.width, widthFactor),
-                      height: barYOffset(section.height, value, maxValue))
-    }
-    
-    internal static func barXOffset(_ total: CGFloat, _ factor: CGFloat) -> CGFloat {
-        return (-(total * factor) / 2) + (total / 2)
-    }
-    
-    internal static func barYOffset(_ total: CGFloat, _ value: Double, _ maxValue: Double) -> CGFloat {
-        return total - (total * divideByZeroProtection(CGFloat.self, value, maxValue))
-    }
-}
